% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CDR_Job_Inten}
\alias{CDR_Job_Inten}
\title{Job Intensities for Various CDR Technologies}
\format{
A data frame with columns:
\describe{
\item{CDR}{Name of the CDR technology.}
\item{min_int}{Minimum job intensity (jobs per tCO2 removed).}
\item{mean_int}{Mean job intensity (jobs per tCO2 removed).}
\item{max_int}{Maximum job intensity (jobs per tCO2 removed).}
\item{units}{Units for job intensity.}
}
}
\usage{
CDR_Job_Inten
}
\description{
A dataset containing job intensities (\code{min_int}, \code{mean_int}, \code{max_int}) for various
Carbon Dioxide Removal (CDR) technologies.
}
\examples{
data(CDR_Job_Inten)
head(CDR_Job_Inten)
}
\keyword{datasets}
